[REF] sale: Separate mail templates by sale state
Prepare files for a simplier wizard based on the fat one.
Changes in action_send_quotation to adapt to the future wizard.
